Rename types module to game_models
